pyproject.toml
[project]
name = "hokuhoku-imomaru-bot"
version = "0.1.0"
description = "X育成ボット「ほくほくいも丸くん」- AWS Serverless Bot"
requires-python = ">=3.12"
dependencies = [
    "aws-cdk-lib>=2.100.0",
    "constructs>=10.0.0",
    "boto3>=1.34.0",
    "Pillow>=10.0.0",
    "requests>=2.31.0",
    "hypothesis>=6.90.0",
    "pytest>=7.4.0",
    "moto>=4.2.0",
    "pytest-cov>=4.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=src --cov-report=term-missing"

[tool.coverage.run]
source = ["src"]
omit = ["tests/*", ".venv/*"]
